+++ worldengine/hdf5_serialization.py
@@ -2,12 +2,14 @@ import numpy
 
 import h5py
 
+
+#fairly certain I should not do this.
 from worldengine.version import __version__
 from worldengine.biome import biome_name_to_index, biome_index_to_name
 from worldengine.model.world import World, Step, Size, GenerationParameters
 
-
 def save_world_to_hdf5(world, filename):
+    
     f = h5py.File(filename, libver='latest', mode='w')
 
     general_grp = f.create_group("general")
