+++ worldengine/simulations/hydrology.py
@@ -1,4 +1,4 @@
-from worldengine.simulations.basic import find_threshold_f
+from simulations.basic import find_threshold_f
 import numpy
