+++ worldengine/simulations/precipitation.py
@@ -2,8 +2,8 @@ import time
 import numpy
 from noise import snoise2
 
-from worldengine.simulations.basic import find_threshold_f
-from worldengine.common import get_verbose
+from simulations.basic import find_threshold_f
+from common import get_verbose
 
 
 class PrecipitationSimulation(object):
